backend/ingestion/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
httpx>=0.27.0
openpyxl>=3.1.0
reportlab>=4.0.0
pdfminer.six>=20221105
python-docx>=0.8.11
sentence-transformers>=2.2.0
numpy>=1.24.0

